requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# MinIO S3 Client
minio==7.2.0

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Utils
python-slugify==8.0.1
httpx==0.26.0

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Authentication
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
